// apps/admin/src/lib/api/generated/models/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.51
 */
import type { AccountOut } from './accountOut';
import type { AssetMetrics } from './assetMetrics';
import type { Totals } from './totals';
import type { Trend } from './trend';

export interface Dashboard {
  account: AccountOut;
  asset_total: number;
  assets: AssetMetrics[];
  end: string;
  last_updated_at: string | null;
  start: string;
  totals: Totals;
  trend: Trend[];
}